Upgrade ListSheet
